renamed class to Tripper. Moved Tripper to a new file tripper/tripper.py. Modified the __init__ and tests to account for the changes
