+++ docs/api_reference/tripper.md
@@ -1,0 +1,3 @@
+# tripper
+
+::: tripper.tripper
